requirements.txt
mysql-connector-python
pandas
scikit-learn
joblib
numpy